exec.c moved to exec module (folder) and separated to child and pathfinding
