+++ src/builtins/echo.c
@@ -47,79 +47,6 @@ static int	print_quotedstr(char *arg)
 	return (EXIT_SUCCESS);
 }
 
-/**This implementation of echo now *writes* every string in the array args by
- * iterating it one by one. This is more useful as control to make sure that
- * echo ONLY goes to stdout. If somehow it fails, it should return a different
- * exit_status (1) instead of the default (0).
- * If the opt is 1, then dont end with a newline, otherwise print it.
- * 		Takes the args array as argument, including the boolean int opt.
- * 		Returns the exit_status 0 as the default, if somehow writing to stdout
- * 			fails, returns exit_status 1.
- */
-/* int	rh_echo(char *args[], int opt)
-{
-	int	i;
-
-	i = 0;
-	while (args[i])
-	{
-		if (i > 0)
-		{
-			if (write(STDOUT_FILENO, " ", 1) == -1)
-			{
-				perror("Echo: write error");
-				return (EXIT_FAILURE);
-			}
-		}
-		if (print_quotedstr(args[i]) != EXIT_SUCCESS)
-		{
-			perror("Echo: write error");
-			return (EXIT_FAILURE);
-		}
-		i++;
-	}
-	if (opt == 0)
-		write(STDOUT_FILENO, "\n", 1);
-	return (EXIT_SUCCESS);
-} */
-
-/* int	rh_echo(char *args[])
-{
-	int	i;
-	int	opt;
-
-	opt = 0;
-	i = 1;
-	if (args[1] && ft_strncmp(args[1], "-n", 2) == 0)
-	{
-		opt = 1;
-		i++;
-	}
-	while (args[i])
-	{
-		if (i > 1)
-		{
-			if (write(STDOUT_FILENO, " ", 1) == -1)
-			{
-				perror("Echo: write error");
-				return (EXIT_FAILURE);
-			}
-		}
-		if (print_quotedstr(args[i]) != EXIT_SUCCESS)
-		{
-			perror("Echo: write error");
-			return (EXIT_FAILURE);
-		}
-		i++;
-
-	}
-	if (opt == 0)
-		write(STDOUT_FILENO, "\n", 1);
-	else
-		write(STDOUT_FILENO, "", 0);
-	return (EXIT_SUCCESS);
-} */
-
 int	check_n(char *arg)
 {
 	int	i;
